refactor: inline linear mla converters
