+++ awex/sharding/param_sharding.py
@@ -289,6 +289,31 @@ class ShardingStrategy:
         return ShardingType.TP_SHARDING, sharding_dim, tp_size
 
 
+class LinearMLAShardingMixin:
+    def get_sharding_strategy(self, parameter_name, **kwargs):
+        if any(
+            key in parameter_name
+            for key in (
+                "attention.kv_a_proj_with_mqa.weight",
+                "attention.q_a_proj.weight",
+                "attention.fused_qkv_a_proj_with_mqa.weight",
+                "attention.kv_a_layernorm.weight",
+                "attention.q_a_layernorm.weight",
+            )
+        ):
+            sharding_dim = get_default_sharding_dim(parameter_name)
+            return ShardingType.NO_SHARDING, sharding_dim, 1
+
+        if "attention.g_norm.weight" in parameter_name:
+            sharding_dim = get_default_sharding_dim(parameter_name)
+            tp_size = self.rank_info.tp_size
+            if tp_size > 1:
+                return ShardingType.TP_SHARDING, sharding_dim, tp_size
+            return ShardingType.NO_SHARDING, sharding_dim, 1
+
+        return super().get_sharding_strategy(parameter_name, **kwargs)
+
+
 def get_sharding_strategy_builder(engine_name: str):
     if engine_name == "sglang":
         from awex.sharding.sglang_sharding import get_sglang_sharding_strategy
